feat(rename): change to poliedro for solutions and projec name
