+++ APP.Eds/APP.Eds/Models/Hose/DispensersResponse.cs
@@ -1,5 +1,4 @@
-﻿using APP.Eds.Models.Dispenser;
-using System.Text.Json.Serialization;
+﻿using System.Text.Json.Serialization;
 
 namespace APP.Eds.Models.Hose
 {
