+++ APP.Eds/APP.Eds/Models/ProductCompartiment/compartimentModelResponse.cs
@@ -1,6 +1,6 @@
 ﻿using Newtonsoft.Json;
 
-namespace APP.Eds.Models.Compartiment;
+namespace APP.Eds.Models.ProductCompartiment;
 
 public class CompartimentModelResponse
 {
