+++ APP.Eds/APP.Eds/Platforms/Android/MainActivity.cs
@@ -2,7 +2,7 @@
 using Android.Content.PM;
 using Android.OS;
 
-namespace APP.Eds
+namespace APP.Eds.Platforms.Android
 {
     [Activity(Theme = "@style/Maui.SplashTheme", MainLauncher = true, LaunchMode = LaunchMode.SingleTop, ConfigurationChanges = ConfigChanges.ScreenSize | ConfigChanges.Orientation | ConfigChanges.UiMode | ConfigChanges.ScreenLayout | ConfigChanges.SmallestScreenSize | ConfigChanges.Density)]
     public class MainActivity : MauiAppCompatActivity
