+++ APP.Eds/APP.Eds/Poliedro.csproj
