+++ APP.Eds/APP.Eds/Services/Court/CourtService.cs
@@ -1,6 +1,5 @@
 ﻿using APP.Eds.Helpers;
 using APP.Eds.Models.Court;
-using APP.Eds.Models.Dispenser;
 using APP.Eds.Models.Eds;
 using APP.Eds.Models.Hose;
 using APP.Eds.Models.Inventory;
