+++ APP.Eds/APP.Eds/Services/ProductCompartiment/ProductCompartimentService.cs
@@ -1,8 +1,6 @@
 ﻿using APP.Eds.Helpers;
-using APP.Eds.Models.Compartiment;
 using APP.Eds.Models.EdsTank;
 using APP.Eds.Models.Islander;
-using APP.Eds.Models.Product;
 using APP.Eds.Models.ProductCompartiment;
 using APP.Eds.Services.Config;
 using System.Collections.ObjectModel;
