+++ APP.Eds/APP.Eds/UsesCases/Product/ProductPostView.xaml.cs
@@ -1,4 +1,3 @@
-using APP.Eds.Models.Product;
 using APP.Eds.Services.Product;
 using APP.Eds.UsesCases.ProductType;
 using APP.Eds.Services.Alert;
@@ -6,6 +5,7 @@ using APP.Eds.Components.PopUp;
 using System.ComponentModel;
 using System.Runtime.CompilerServices;
 using System.Windows.Input;
+using APP.Eds.Models.ProductCompartiment;
 
 
 namespace APP.Eds.UsesCases.Product;
