+++ APP.Eds/Poliedro.sln
@@ -3,7 +3,7 @@ Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio Version 17
 VisualStudioVersion = 17.12.35521.163
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "APP.Eds", "APP.Eds\APP.Eds.csproj", "{15971DFA-F3EE-4791-A89C-7887529D877A}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Poliedro", "APP.Eds\Poliedro.csproj", "{15971DFA-F3EE-4791-A89C-7887529D877A}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "TestAutomation", "TestAutomation", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
 EndProject
